Backend/story_project/requirements.txt
django
djangorestframework
boto3
botocore
python-dotenv
openai
django-cors-headers
gunicorn